Agregada la clase ReceiptShould en el paquete Tests. Se cambió el tipo de Commission en Provider de long a double. Los tests de las clases ProviderShould, TextFieldShould, DateFieldShould y NumberFieldShould fueron refactoreados para que ya no llevaran "Test" sobre el final del nombre.

+++ TodoPagos/Domain/Provider.cs
@@ -5,7 +5,7 @@ namespace Domain
 {
     public class Provider
     {
-        public long Commission { get; set; }
+        public double Commission { get; set; }
 
         public string Name { get; set; }
